src/ui/requirements.txt
## Minimal dependencies for POI Recommender Streamlit UI
## Using pre-built wheels to avoid compilation

# Core UI framework
streamlit>=1.28.0

# Data & visualization
plotly>=5.18.0
folium>=0.14.0
streamlit-folium>=0.14.0
matplotlib>=3.7.0
wordcloud>=1.9.0
pillow>=10.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# Google Cloud integration
google-cloud-storage>=2.10.0
cloud-sql-python-connector>=1.4.0
sqlalchemy>=2.0.0
pg8000>=1.30.0

# Note: numpy, pandas, torch, duckdb are already available 
# as dependencies of the main project (src/models, src/data)
# If you need to use the UI standalone, uncomment below:

# pip install numpy pandas duckdb torch --index-url https://download.pytorch.org/whl/cpu
# pytest==7.4.3
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
